Fixed HUE wheel.

+++ firmware/src/apps/light_dimmer/light_dimmer.h
@@ -56,12 +56,12 @@ private:
     bool color_set = false;
 
     uint16_t app_hue_position = 0;
-    uint8_t current_brightness = 0;
+    uint16_t current_brightness = 0;
     bool is_on = false;
 
     bool first_run = false;
 
-    uint16_t calculateAppHuePosition(uint16_t position);
+    int8_t calculateAppHuePosition(int8_t position);
 
     lv_obj_t *mask_img;
     lv_obj_t *dimmer_screen;
